C64997: Asterisk issue: asterisk or italics?
Hello @anirudhcavale,
Localization team has reported source content issue that causes localized version to have broken/different format compared to en-us version.
Description: Isolated asterisks are shown on live branch. Are they supposed to be left as-is (if so consider using backslashes to scape them) or was it a translation mistake and opening italics tags (*) were missing?
Please review and merge the proposed file change to fix to target versions. If you make related fix in another PR  then share your PR number so we can confirm and close this PR.
Many thanks in advance.