Fix client/browser-side trackTrace documentation
trackTrace does not - at least in the browser - take 3 separate parameters. It takes a single parameter, an object, with properties `message` and `properties`. I left the documentation in for `severityLevel` but I am unconvinced that the existing documentation on providing it is accurate, as I was unable to use it in my testing.