Fixed FastAPI sample code memory leak
There is a memory leak in the FastAPI / Tracer sample code. In the code, the Tracer is created on every request, and that leaks hundreds of KB of memory for each request. Each AzureExporer instantiation will spin up worker threads and will remain allocated even after the request is over.

We should only instantiate the Tracer once outside of the middleware function.

This bug was found separately in the opencensus repo.
https://github.com/census-instrumentation/opencensus-python/issues/1020#issuecomment-1022549110