Clarifying Retention Setting For Classic Data
Trying to clarify that the Retention setting in App Insights ONLY controls the retention setting on telemetry saved in the classic storage. To change retention of classic data use retention setting in App Insights, to change retention of workspaced backed data, change retention setting in Log Analytics.  For customers with data still in both classic and workspace, understanding this difference matters.