Update data-retention-archive.md
I believe the limits in the current version of docs are wrong. Total retention cannot be set to 2,555 days, because it's failing with "The specified total retention value is invalid. The value beyond two years is restricted to full years. Allowed values are: [4-730], 1095, 1460, 1826, 2191, 2556 days.". And this value range applies to totalRetentionInDays, not to retentionInDays (as it is contradictory to "You can keep data in interactive retention between 4 and 730 days." stated above)