Regional endpoints only support TLS 1.2
As Breeze 2.0 rolls out the regional endpoints will only support TLS 1.2 clients. We need to get this documented, similar to how we documented this for Live Metrics endpoint. As part of this change, adding a column to the IP list table that helps customers identify which of those endpoints are considered 'global' versus being 'regional'.