Correcting dashboard creation statement
The default AI dashboard does NOT get create at component resource creation. It ONLY gets created if it does not already exist and the customer selects the Application Dashboard button from the Overview pane of App Insights. This new modification to the content makes it clearer and also completes the full picture in the rest of the paragraph  where we have you delete that default dashboard and get it recreated with the updated resource URI.