notify that table names changed
Hey team, customer called in 2305010010001837 and was trying to adjust retention settings on their app insights tables. The portal UX navigated the customer to this App Insights paragraph however we do not inform the customer that the table names change as you go over to LA. I want to add this tip so that customers will realize that if they want to change retention settings for customMetrics table then they have to adjust that setting for the AppMetrics table in Log Analytics. Hoping this edit here does the trick.