Adding billing FAQ for snapshots
We don't have any docs about if customers are charged for these snapshot dump files. It isn't clear if we charge for ingestion or storage, which we do neither. There are no charges for using snapshots or holding onto customer files in our storage so just want to get this documented. I think this TSG is a good spot.