Support for Unicode on release annotations
We have had two customers call in to open support cases because the sample script on this page was choking when customers supplied unicode characters as part of the release name or release properties json body payload. This update adds support for unicode characters to allow those annotations to get created without failures.