azure-monitor: simplify relabel_configs examples
The intent of the `relabl_configs` examples in this document is to apply
a new label to all metrics. The `source_labels` field is not needed, as
the re-labeling does not use the values from `__address__`, and
specifying `source_labels` is not necessary to select all metrics for
re-labeling. We can simply omit this from the suggested configuration.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>