Update connection-strings.md
Updates to the Connection String formats have been made in Early 2024.
The public docs have not been updated yet so just make the changes. Worked on SR#2502070040013430 where cx asked what the app ID was as he does not recall seeing it in the past.
Reviewed our internal comms and explained to the cx that we made the changes in 2024. cx asked if this is documented publicly.