Adding clarity for classic App Insights SDK API instructions
One of my customers got confused about this page because they're using the classic API in their code and was not sure which steps to follow. This is how I'm proposing to add some clarity but feel to format the article how you think is better. We just need to make sure that customers still using the classic API know which steps to follow to add Entra ID authentication.