Update advisor-overview.md
While working on a case, a customer raised several concerns. I reached out to the PG team to gather the necessary information and conveyed it to the customer. Additionally, I updated the TSG with this new information, which was not previously available in any existing TSGs or public wiki. This update will be beneficial for my colleagues in future investigations.