Update permissions details for log alert rules
Had a customer recently and I'm sending this change based on the discoveries made after discussing with the Alerts PG. This gap was missing here. Clarified the permissions requirements for alert rules, especially regarding managed identities and Log Analytics workspaces. Added an example to illustrate the access needed for resource-centric alerts.