Fix sampler type naming in OpenTelemetry config
After running into issues with the fixed percentage not working, I started looking at the PR which made these Environment variables possible, see https://github.com/Azure/azure-sdk-for-net/pull/52720/files. I found out the naming is with an underscore, not a dot, causing the sampling to not be applied and the exporter to ignore the values given.